requirements.txt
# Core deep learning
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
deepspeed>=0.12.0

# NVIDIA Transformer Engine (FP4/FP8 quantization on B200)
transformer-engine[pytorch]>=1.5.0

# Data processing
datasets>=2.16.0
pyarrow>=14.0.0
pandas>=2.0.0
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Tokenization
tokenizers>=0.15.0

# Evaluation
rouge-score>=0.1.2
evaluate>=0.4.1

# Monitoring
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0

# Flash Attention (requires compilation)
flash-attn>=2.5.0

# Testing
pytest>=7.4.0
